feat(local): consolidar comunicacao e reconciliacao

+++ app/Console/Commands/SetupCatalogoE2EFixturesCommand.php
@@ -4,6 +4,7 @@ namespace App\Console\Commands;
 
 use Illuminate\Console\Command;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Schema;
 
 class SetupCatalogoE2EFixturesCommand extends Command
 {
@@ -80,7 +81,9 @@ class SetupCatalogoE2EFixturesCommand extends Command
 
                 DB::table('estoque_movimentacoes')->whereIn('pedido_id', $pedidoIds)->delete();
                 DB::table('estoque_reservas')->whereIn('pedido_id', $pedidoIds)->delete();
-                DB::table('pedido_status_historicos')->whereIn('pedido_id', $pedidoIds)->delete();
+                if (Schema::hasTable('pedido_status_historicos')) {
+                    DB::table('pedido_status_historicos')->whereIn('pedido_id', $pedidoIds)->delete();
+                }
 
                 if ($pedidoItemIds !== []) {
                     DB::table('pedido_itens')->whereIn('id', $pedidoItemIds)->delete();
